+++ .github/workflows/release.yml
@@ -7,6 +7,11 @@ on:
     tags:
       - "v*"
   workflow_dispatch:
+    inputs:
+      build_official_preload:
+        description: 'Build and publish official preload package'
+        type: boolean
+        default: true
 
 jobs:
   release:
@@ -36,79 +41,15 @@ jobs:
       - name: Build Engine
         run: pnpm b:all
 
-      - name: Cache Engine Package
-        uses: actions/cache@v3
-        with:
-          path: ./packages/galacean
-          key: ${{ runner.os }}-engine-${{ github.sha }}
-
-      - name: Checkout Sub-repositories
+      - name: Make scripts executable
         run: |
-          git clone https://github.com/galacean/engine-toolkit.git
-          git clone https://github.com/galacean/engine-lottie.git
-          git clone https://github.com/galacean/engine-spine.git -b 4.2
+          chmod +x ./scripts/build-editor-preload.js
+          chmod +x ./scripts/build-official-preload.js
 
-      - name: Checkout editor repository
-        uses: actions/checkout@v4
-        with:
-          repository: galacean/editor
-          submodules: true
-          path: editor
-          token: ${{ secrets.CLONE_EDITOR_TOKEN }}
-
-      - name: Install and Build for editor
-        working-directory: ./editor
-        run: |
-          pnpm install --no-frozen-lockfile
-          pnpm build:decorators
-
-      - name: Install and Link Engine and Build for Toolkit
-        working-directory: ./engine-toolkit
-        run: |
-          pnpm install
-          pnpm link ../packages/galacean
-          pnpm b:all
-
-      - name: Install and Link Engine and Build for Lottie
-        working-directory: ./engine-lottie
-        run: |
-          pnpm install --ignore-workspace
-          pnpm link ../packages/galacean
-          pnpm build
-
-      - name: Install and Link Engine and Build for Spine
-        working-directory: ./engine-spine
-        run: |
-          pnpm install --ignore-workspace --frozen-lockfile=false
-          pnpm link ../packages/galacean
-          pnpm build
-
-      - name: Structure Temp Directory
-        run: |
-          mkdir -p ${{ github.workspace }}/temp
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine
-          cp -r ${{ github.workspace }}/packages/galacean/dist ${{ github.workspace }}/temp/@galacean/engine
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine-xr
-          cp -r ${{ github.workspace }}/packages/xr/dist ${{ github.workspace }}/temp/@galacean/engine-xr
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine-ui
-          cp -r ${{ github.workspace }}/packages/ui/dist ${{ github.workspace }}/temp/@galacean/engine-ui
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine-shaderlab
-          cp -r ${{ github.workspace }}/packages/shader-lab/dist ${{ github.workspace }}/temp/@galacean/engine-shaderlab
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine-shader-shaderlab
-          cp -r ${{ github.workspace }}/packages/shader-shaderlab/dist ${{ github.workspace }}/temp/@galacean/engine-shader-shaderlab
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine-physics-lite
-          cp -r ${{ github.workspace }}/packages/physics-lite/dist ${{ github.workspace }}/temp/@galacean/engine-physics-lite
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine-physics-physx
-          cp -r ${{ github.workspace }}/packages/physics-physx/dist ${{ github.workspace }}/temp/@galacean/engine-physics-physx
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine-toolkit
-          cp -r ${{ github.workspace }}/engine-toolkit/packages/galacean-engine-toolkit/dist/umd ${{ github.workspace }}/temp/@galacean/engine-toolkit
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine-lottie
-          cp -r ${{ github.workspace }}/engine-lottie/dist ${{ github.workspace }}/temp/@galacean/engine-lottie
-          mkdir -p ${{ github.workspace }}/temp/@galacean/engine-spine
-          cp -r ${{ github.workspace }}/engine-spine/dist ${{ github.workspace }}/temp/@galacean/engine-spine
-          mkdir -p ${{ github.workspace }}/temp/@galacean/editor-decorators
-          cp -r ${{ github.workspace }}/editor/packages/decorators/dist ${{ github.workspace }}/temp/@galacean/editor-decorators
-          find ${{ github.workspace }}/temp
+      # Build and publish official preload package
+      - name: Build official preload package
+        if: ${{ github.event.inputs.build_official_preload == '' || github.event.inputs.build_official_preload == 'true' }}
+        run: ./scripts/build-official-preload.js
 
       - name: Release engine packages and Sync to CDN
         uses: galacean/publish@main
@@ -119,3 +60,25 @@ jobs:
           OASISBE_UPLOAD_URL: https://oasisbe.alipay.com/api/file/no-auth/crypto/upload
           OASISBE_REQUEST_HEADER: ${{secrets.OASISBE_REQUEST_HEADER}}
           OASISBE_PUBLIC_KEY: ${{secrets.OASISBE_PUBLIC_KEY}}
+
+      - name: Release official preload package and Sync to CDN
+        if: ${{ github.event.inputs.build_official_preload == '' || github.event.inputs.build_official_preload == 'true' }}
+        uses: galacean/publish@main
+        with:
+          publish: false
+          packages: |
+            editor-preload-official
+        env:
+          NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
+          NPM_CONFIG_PROVENANCE: true
+          OASISBE_UPLOAD_URL: https://oasisbe.alipay.com/api/file/no-auth/crypto/upload
+          OASISBE_REQUEST_HEADER: ${{secrets.OASISBE_REQUEST_HEADER}}
+          OASISBE_PUBLIC_KEY: ${{secrets.OASISBE_PUBLIC_KEY}}
+          
+      - name: Upload official preload artifact
+        if: ${{ github.event.inputs.build_official_preload == '' || github.event.inputs.build_official_preload == 'true' }}
+        uses: actions/upload-artifact@v4
+        with:
+          name: editor-preload-official
+          path: editor-preload-official/dist/browser.js
+          retention-days: 7
